--- rllm/experimental/fireworks/__init__.py
@@ -1,9 +1,0 @@
-from rllm.experimental.fireworks.fireworks_backend import FireworksBackend
-from rllm.experimental.fireworks.fireworks_launcher import FireworksTrainerLauncher
-from rllm.experimental.fireworks.fireworks_policy_trainer import FireworksPolicyTrainer
-
-__all__ = [
-    "FireworksBackend",
-    "FireworksTrainerLauncher",
-    "FireworksPolicyTrainer",
-]
