+++ rllm/trainer/fireworks/__init__.py
@@ -1,0 +1,9 @@
+from rllm.trainer.fireworks.fireworks_backend import FireworksBackend
+from rllm.trainer.fireworks.fireworks_launcher import FireworksTrainerLauncher
+from rllm.trainer.fireworks.fireworks_policy_trainer import FireworksPolicyTrainer
+
+__all__ = [
+    "FireworksBackend",
+    "FireworksTrainerLauncher",
+    "FireworksPolicyTrainer",
+]
